requirements.txt
spotipy==2.25.1
requests==2.32.3
pillow==11.1.0
